Rewrote GUI in Pyside
As per the last commit comment, there were issues with using the OpenCV video display window. For this reason I've rewritten everything in Pyside. This solved the problems with keypresses, and as a further advantage, improved the general software architecture of the program - a huge benefit for the next few tutorials